refactor(RandomNumGenerator): domain 로직을 수행하기에 domain 패키지로 이동
