+++ src/main/java/lotto/domain/RandomNumGenerator.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import lotto.system.Constant;
