+++ src/test/java/lotto/domain/RandomNumGeneratorTest.java
@@ -1,9 +1,9 @@
-package lotto.util;
+package lotto.domain;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.util.RandomNumGenerator;
+import lotto.domain.RandomNumGenerator;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
